Day-10/Final Hackathon/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.7.2
langgraph==0.1.9
langchain==0.2.11
langchain-core==0.2.23
langchain-google-genai==1.0.7
chromadb==0.4.18
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pydantic==2.7.4
pydantic-settings==2.1.0
httpx==0.25.2
requests==2.31.0 